requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
torch==2.0.1
torchvision==0.15.2
python-multipart==0.0.5
jinja2==3.0.1
aiofiles==0.7.0
numpy==1.21.2
python-socketio==5.5.0
plotly==5.3.1 